requirements.txt
python-telegram-bot==22.6
httpx==0.27.2
Flask==3.1.3
urllib3<3.0
setuptools>=68.0.0
